+++ Samples~/CoreStateBridgeSample/CoreStateBridgeSampleRaycastController.cs
@@ -1,8 +1,0 @@
-using Deucarian.ObjectSelection;
-
-namespace Deucarian.ObjectSelection.CoreStateBridge.Samples
-{
-    public sealed class CoreStateBridgeSampleRaycastController : RaycastSelectionController<string>
-    {
-    }
-}