+++ Samples~/CoreStateIntegrationSample/CoreStateIntegrationSampleRaycastController.cs
@@ -1,0 +1,8 @@
+using Deucarian.ObjectSelection;
+
+namespace Deucarian.ObjectSelection.CoreStateIntegration.Samples
+{
+    public sealed class CoreStateIntegrationSampleRaycastController : RaycastSelectionController<string>
+    {
+    }
+}